requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
sqlparse==0.5.3
django-environ==0.10.0
python-dotenv
psycopg2-binary
django-import-export
openpyxl>=3.1.2
